+++ src/main/java/com/likelion/RePlay/domain/info/entity/InfoSubmit.java
@@ -1,0 +1,30 @@
+package com.likelion.RePlay.domain.info.entity;
+
+import com.likelion.RePlay.domain.user.entity.User;
+import com.likelion.RePlay.global.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name="INFO_SUBMIT")
+public class InfoSubmit extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="INFO_SUBMIT_ID")
+    private Long infoSubmitId;
+
+    @Column(name = "TITLE")
+    private String title;
+
+    @Column(name = "CONTENT")
+    private String content;
+
+    @ManyToOne
+    @JoinColumn(name = "USER_ID")
+    private User user; // 투고 작성 회원
+}
