+++ src/main/java/com/likelion/RePlay/domain/learning/entity/LearningComment.java
@@ -1,0 +1,31 @@
+package com.likelion.RePlay.domain.learning.entity;
+
+import com.likelion.RePlay.domain.user.entity.User;
+import com.likelion.RePlay.global.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name="LEARNING_COMMENT")
+public class LearningComment extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="Learning_COMMENT_ID")
+    private Long LearningCommentId;
+
+    @Column(name="CONTENT")
+    private Long content;
+
+    @ManyToOne
+    @JoinColumn(name = "USER_ID")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "LEARNING_ID")
+    private Learning learning;
+}
