+++ src/main/java/com/likelion/RePlay/domain/playing/entity/PlayingScrap.java
@@ -1,0 +1,28 @@
+package com.likelion.RePlay.domain.playing.entity;
+
+import com.likelion.RePlay.domain.user.entity.User;
+import com.likelion.RePlay.global.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name="PLAYING_SCRAP")
+public class PlayingScrap extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="PLAYING_SCARP_ID")
+    private Long playingScrapId;
+
+    @ManyToOne
+    @JoinColumn(name = "USER_ID")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "PLAYING_ID")
+    private Playing playing;
+}
