--- src/main/java/com/likelion/RePlay/entity/playing/PlayingApply.java
@@ -1,28 +1,0 @@
-package com.likelion.RePlay.entity.playing;
-
-import com.likelion.RePlay.entity.User;
-import com.likelion.RePlay.util.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Entity
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name="PLAYING_APPLY")
-public class PlayingApply extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="PLAYING_APPLY_ID")
-    private Long playingApplyId;
-
-    @ManyToOne
-    @JoinColumn(name = "USER_ID")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "PLAYING_ID")
-    private Playing playing;
-}
