--- src/main/java/com/likelion/RePlay/entity/playing/PlayingScrap.java
@@ -1,28 +1,0 @@
-package com.likelion.RePlay.entity.playing;
-
-import com.likelion.RePlay.entity.User;
-import com.likelion.RePlay.util.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Entity
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name="PLAYING_SCRAP")
-public class PlayingScrap extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="PLAYING_SCARP_ID")
-    private Long playingScrapId;
-
-    @ManyToOne
-    @JoinColumn(name = "USER_ID")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "PLAYING_ID")
-    private Playing playing;
-}
