--- src/main/java/com/likelion/RePlay/enums/Category.java
@@ -1,5 +1,0 @@
-package com.likelion.RePlay.enums;
-
-public enum Category {
-    IMPROVEMENT // 자기계발, LANGUAGE, SNS, BOOK, CULTURE, PICNIC // 나들이, TRIP, EXERCISE, RESTAURANT, ENTERTAINMENT, COOKING, ETC
-}
