--- src/main/java/com/likelion/RePlay/enums/IsCompleted.java
@@ -1,5 +1,0 @@
-package com.likelion.RePlay.enums;
-
-public enum IsCompleted {
-    TRUE, FALSE
-}
