--- src/main/java/com/likelion/RePlay/enums/IsRecruit.java
@@ -1,5 +1,0 @@
-package com.likelion.RePlay.enums;
-
-public enum IsRecruit {
-    TRUE, FALSE
-}
