--- src/main/java/com/likelion/RePlay/enums/Role.java
@@ -1,5 +1,0 @@
-package com.likelion.RePlay.enums;
-
-public enum Role {
-    USER, ADMIN
-}
