--- src/main/java/com/likelion/RePlay/enums/State.java
@@ -1,6 +1,0 @@
-package com.likelion.RePlay.enums;
-
-public enum State {
-    SEOUL, PUSAN, DAEGU, INCHEON, GWANGJU, DAEJEON, ULSAN, SEJONG,
-    GEYONGGI, KANGWON, CHUNGBUK,CHUNGNAM, JEONBUK, JEONNAM, KYUNGBUK, KYUNGNAM, JEJU
-}
