+++ src/main/java/com/likelion/RePlay/global/enums/Category.java
@@ -1,0 +1,5 @@
+package com.likelion.RePlay.global.enums;
+
+public enum Category {
+    IMPROVEMENT // 자기계발, LANGUAGE, SNS, BOOK, CULTURE, PICNIC // 나들이, TRIP, EXERCISE, RESTAURANT, ENTERTAINMENT, COOKING, ETC
+}
