+++ src/main/java/com/likelion/RePlay/global/enums/District.java
@@ -1,0 +1,27 @@
+package com.likelion.RePlay.global.enums;
+
+public enum District {
+    ALL, // 전체 혹은 시에서 세종 선택시
+    GANGNAM, GANGDONG, GANGBUK, GANGSEO, GWANAK, GWANGJIN, GURO, GEUMCHEON, // 서울
+    NOWON, DOBONG, DONGDAEMUN, DONGJAK, MAPO, SEODAEMUN, SEOCHO, SEONGDONG, // 서울
+    SEONGBUK, SONGPA, YANGCHEON, YEONGDEUNGPO, YONGSAN, EUNPYEONG, JONGROM, JUNG, JUNGRANG, // 서울
+
+    GEUNGJUNG, NAM, DONG, DONGRAE, PUSANJIN, SASANG, SAHA, SEO, SUYOUNG, YEONJE, YOUNGDO, HAEUNDAE,
+
+    DALSEO, DALSEONG, BUK, SUSEONG,
+    GYEYANG, NAMDONG, MICHUHOL, BUPYEONG, YEONSU,
+    GWANGSAN,
+    DAEDEOK, YUSUNG,
+    GOYANG, GWACHEON, GWANGMYEONG, GWANGJU, GURI, GUNPO, GIMPO, NAMYANGJU, // 경기
+    DONGDOOCHEON, BUCHEON, SUNGNAM, SUWON, SIHEUNG, ANSAN, ANSUNG, ANYANG, YANGJU, // 경기
+    YEOJU, OSAN, YONGIN, UIWANG, UIJUNGBU, ICHEON, PAJU, PYEONGTAEK, POCHEON, HANAM, HWASUNG, // 경기
+
+    GANGNEUNG, DONGHAE, SAMCHUCK, SOKCHO, WONJU, CHUNCHEON, TAEBAEK,
+    JECHEON, CHEONGJU, CHUNGJU, GYERYONG, GONGJU, NONSAN, DANGJIN, BORYEONG, SEOSAN, ASAN, CHEONAN,
+    GUNSAN, GIMJE, NAMWON, IKSAN, JEONJU, JUNGEUP,
+    GWANGYANG, NAJU, MOKPO, SUNCHEON, YEOSU,
+    GYUNGSAN, GYUNGJU, GUMI, GIMCHEON, MUNGYUNG, SANGJU, ANDONG, YOUNGJU, POHANG,
+    GEOJE, GIMHAE, MASAN, MILYANG, SACHEON, YANGSAN, JINJU, CHANGWON, TONGYEONG,
+    SEOGUIPO, JEJU
+
+}
