+++ src/main/java/com/likelion/RePlay/global/enums/IsCompleted.java
@@ -1,0 +1,5 @@
+package com.likelion.RePlay.global.enums;
+
+public enum IsCompleted {
+    TRUE, FALSE
+}
