+++ src/main/java/com/likelion/RePlay/global/enums/IsRecruit.java
@@ -1,0 +1,5 @@
+package com.likelion.RePlay.global.enums;
+
+public enum IsRecruit {
+    TRUE, FALSE
+}
