--- src/main/java/com/likelion/RePlay/util/entity/BaseEntity.java
@@ -1,25 +1,0 @@
-package com.likelion.RePlay.util.entity;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(name="created_at")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(name="updated_at")
-    private LocalDateTime updatedAt;
-}
