--- src/main/java/com/likelion/RePlay/util/exception/CustomExceptionHandler.java
@@ -1,41 +1,0 @@
-package com.likelion.RePlay.util.exception;
-
-import com.likelion.RePlay.util.response.CustomAPIResponse;
-import jakarta.validation.ConstraintViolation;
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.boot.web.servlet.error.ErrorController;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.util.stream.Collectors;
-
-@ControllerAdvice
-public class CustomExceptionHandler implements ErrorController {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<CustomAPIResponse<?>> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
-        String errorMessage = e.getBindingResult().getAllErrors().stream()
-                .map(DefaultMessageSourceResolvable::getDefaultMessage)
-                .collect(Collectors.joining("; "));
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(CustomAPIResponse.createFailWithout(HttpStatus.BAD_REQUEST.value(), errorMessage));
-    }
-
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<CustomAPIResponse<?>> handleConstraintViolationException(ConstraintViolationException e) {
-        String errorMessage = e.getConstraintViolations().stream()
-                .map(ConstraintViolation::getMessage)
-                .collect(Collectors.joining("; "));
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(CustomAPIResponse.createFailWithout(HttpStatus.BAD_REQUEST.value(), errorMessage));
-    }
-
-}
